# server/common/api.yml
swagger: "2.0"
info:
  version: 1.0.0
  title: x-commerce
  description: My cool app
basePath: /api/v1
...
consumes:
  - application/json
produces:
  - application/json
definitions:
...
  UserBody:
    type: object
    tilte: user information
    properties:
      username:
        type: string
        unique: true
        example: "admin"
      firstName:
        type: string
      lastName:
        type: string
      email:
        type: string
      password:
        type: string
        example: "admin"
      phone:
        type: string
  ProductBody:
    type: object
    title: product information
    properties:
      category:
        type: string
        example: aa
      name: 
        type: string
        example: Cake
      photoUrls:
        type: array
        items:
          type: string
          example: aaaa
      price:
        type: number
      status:
        type: string
        description: cake status in store
        enum:
          - available
          - pending
          - sold
...
paths:
...
  /user/{userId}:
    get:
      tags:
        - Users
      summary: Get users by Name
      description:
      parameters:
        - name: userId         
          in: path
          description: The name that need to be fetched. User user1 for testing
          require: true
          type: string
      responses:
        200:
          description: successfull get user
        400:
          description: invalid username supplied
        404:
          description: user not found  

    put:
      tags:
        - Users
      summary: updated user
      description: This can only be done by logged in user
      parameters:
        - name: userId
          in: path
          description: name that need to be updated
          required: true
          type: string
        - name: body
          in: body
          description: Updated user object
          required: true
          schema: 
            $ref: "#/definitions/UserBody"
      responses:
        400:
          description: Invalid user supplied
        404:
          description: user not found
    
    delete:
      tags:
        - Users
      summary: Delete user
      description: This can only be done by logged in user
      parameters:
        - name: userId
          in: path
          description: The name that need to be deleted
          required: true
          type: string
      response:
        400: 
          description: Invalid user supplied
        404:
          description: user not found
...
  /product/{productId}:
    get:
      tags:
        - Products
      summary: Find Product by id
      description: Return a single Product
      parameters:
        - name: productId
          in: path
          description: ID of product to return
          required: true
          type: string
      responses:
        200: 
          description: Successful operation
          schema:
            $ref: "#/definitions/ProductBody"
        400:
          description: Invalid ID supplied
        404:
          description: Product not found
      
    put:
      tags:
        - Products
      summary: Update a product in store with formdata
      description: ""
      parameters:
        - name: productId
          in: path
          description: ID of product that needs to updated
          required: true
          type: string
        - name: product
          in: body
          description: Product object that needs to be added to the store
          required: true
          schema:
            $ref: "#/definitions/ProductBody"
       
      responses: 
        405:
          description: Invalid input

    delete:
      tags:
        - Products
      summary: Delete a product
      description: ""
      parameters:
        - name: productId
          in: path
          description: ProductId to delete
          required: true
          type: string
        
      responses:
        400:
          description: Invalid ID supplied
        404:
          description: Product not found
...